Only send 'Dataset failed to process' Slack notification after last retry
