+++ backend/corpora/upload_failures/app.py
@@ -2,10 +2,13 @@ import os
 
 from backend.corpora.common.utils.aws import delete_many_from_s3
 from backend.corpora.upload_failures.upload import update_dataset_processing_status_to_failed
+from backend.corpora.upload_failures.slack import notify_slack_failure
 
 
 def handle_failure(event, context):
     dataset_uuid = event["dataset_uuid"]
+    if os.getenv("DEPLOYMENT_STAGE") == "prod":
+        notify_slack_failure(dataset_uuid)
     object_key = os.path.join(os.environ.get("REMOTE_DEV_PREFIX", ""), dataset_uuid).strip("/")
     delete_many_from_s3(os.environ["ARTIFACT_BUCKET"], object_key)
     cellxgene_bucket = f"hosted-cellxgene-{os.environ['DEPLOYMENT_STAGE']}"
